requirements.txt
##
#Django
##
Django==1.9
django-url-tracker-fork==0.2.0
futures==3.0.3
django-environ==0.4.0

##
#Caching
##
django-clear-cache==0.3
django-redis==4.3.0
django-heroku-memcacheify==0.8
pylibmc==1.5.0
django-dumper==0.2.7

##
#Queue
##
celery==3.1.19

##
#Images
##
django-simpleimages==1.3.3

##
#Templates
##
django-sekizai==0.9.0

##
#Admin
##
django-admin-sortable2==0.6.2
django-autocomplete-light==2.2.10

##
#Database
##
dj-database-url==0.3.0
psycopg2==2.6.1

##
#Storage
##
django-s3-storage==0.9.8

##
#Static
##
whitenoise==2.0.6

##
#Logging
##
raven==5.9.0

##
#Tracing
##
newrelic==2.60.0.46

##
#Production
##
gunicorn==19.4.1

##
#Development
##
django-debug-toolbar==1.4
django-devserver==0.8.0

##
#Testing
##
factory-boy==2.6.0
django-webtest==1.7.8
WebTest==2.0.18
pytest-django==2.9.1
pytest-env==0.6.0
pytest-xdist==1.13.1
pytest-flake8==0.1
pytest-cov==2.2.0
django-fake-model==0.1.3
